+++ test/unit/instructions/test_tag_parser.rb
@@ -16,17 +16,14 @@ class TagParserTest < Minitest::Test
     assert_equal 'default', tested.preset_name
   end
 
-  # given preset name
   def test_given_preset
     assert_equal 'preset', tested('preset img.jpg').preset_name
   end
 
-  # base image
   def test_base_image
     assert_equal 'img.jpg', tested.source_names.first
   end
 
-  # media images
   def test_media_images
     params = 'img.jpg mobile: mobile.jpg'
     assert_equal 'mobile.jpg', tested(params).source_names[1]
@@ -50,10 +47,37 @@ class TagParserTest < Minitest::Test
     assert_equal 'white space.jpg', tested(params).source_names.first
   end
 
-  # leftovers
+  def test_gravity
+    params = 'img.jpg southeast 3:2 mobile: img.jpg 4:3 center'
+    correct = {
+      nil => 'southeast',
+      'mobile' => 'center'
+    }
+
+    assert_equal correct, tested(params).gravities
+  end
+
+  def test_geometry
+    params = 'img.jpg southeast 3:2 mobile: img.jpg 4:3-10+20 center'
+    correct = {
+      nil => '3:2',
+      'mobile' => '4:3-10+20'
+    }
+
+    assert_equal correct, tested(params).geometries
+  end
+
   def test_leftovers
-    correct = ['--picture', 'class="some class"']
-    params = 'img.jpg mobile: mobile.jpg --picture class="some class"'
+    correct = ['class="implicit"', '--picture', 'class="some class"']
+    params =
+      'img.jpg mobile: mobile.jpg 3:2 center class="implicit"'\
+      ' --picture class="some class"'
     assert_equal correct, tested(params).leftovers
+  end
+
+  def test_bad_arg
+    assert_raises ArgumentError do
+      tested 'img.jpg asdf.jpg 3:2'
+    end
   end
 end