+++ test/unit/srcsets/test_helper_srcsets.rb
@@ -19,20 +19,22 @@ module TestHelperSrcset
     [100, 150, 200, 300].each { |i| stub_generated(i, gstub(i)) }
   end
 
-  def gstub(width)
+  def gstub(width, exists: false)
     GeneratedImageStub.new(
       name: "img-#{width}-aaaaa.jpg",
       uri: "/img-#{width}-aaaaa.jpg",
       width: width,
       format: 'jpg',
-      exists?: false
+      exists?: exists
     )
   end
 
-  def stub_generated(width, returns, format = 'original')
+  def stub_generated(
+    width, returns, format = 'original', crop: nil, gravity: 'center'
+  )
     GeneratedImage.stubs(:new)
                   .with(source_file: @source_image, width: width,
-                        format: format)
+                        format: format, crop: crop, gravity: gravity)
                   .returns(returns)
   end
 end