+++ test/unit/srcsets/test_pixel_ratio.rb
@@ -9,7 +9,7 @@ class TestSrcsetPixelRatio < Minitest::Test
     PictureTag.stubs(preset: {
                        'pixel_ratios' => [1, 1.5, 2, 3],
                        'base_width' => 100
-                     })
+                     }, crop: nil, gravity: 'center')
 
     @tested = Srcsets::PixelRatio.new(@source_image, 'original')
   end